Plugin: XAPI: Move code of statement attachments for portfolio items - refs BT#18201
